propagate to skysub

+++ pypeit/core/combine.py
@@ -177,7 +177,7 @@ def weighted_combine(weights, sci_list, var_list, inmask_stack,
 
     nused = np.sum(mask_stack, axis=0)
     weights_stack = broadcast_weights(weights, shape)
-    weights_mask_stack = weights_stack*mask_stack
+    weights_mask_stack = weights_stack*mask_stack.astype(float)
 
     weights_sum = np.sum(weights_mask_stack, axis=0)
     inv_w_sum = 1./(weights_sum + (weights_sum == 0.0))
